+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add on-demand password chaining
 - add particle effects when on elevator
 - add quantum elevator block
+- add support for portable elevator pad